+++ NEWS.md
@@ -22,6 +22,7 @@
 - fix `calculateOverlap()`when raster aggregation finds polygons with no values
 - fix `createGiottoPolygon()` dispatch on `character` so that it can access poly cleanup params
 - fix incorrect `giottoInstructions` class in older objects now possible via `updateGiottoObject()`
+- Remove imports on deprecated `convHull()`, `minRect()`, `minCircle()`, in favor of `hull()` usage [#1153](https://github.com/drieslab/Giotto/issues/1153) by demographix
 
 ## changes
 - `remove_background_poly` now defaults to `TRUE` during polygon ingestion